moved application table query to queries.ts
